[Docs] A bit more generous info
I would suggest to either provide more (all?) the basic steps that are needed to setup TYPO3 along with headless, or simply and quickly point to the `headless` repo right away (like it is done in the last sentence). IMO, the second option makes more sense, because there is no need to talk about composer nor a TYPo3 extension, when documenting the Nuxt frontend.